Add XMPMeta examples
Add doc tests, but only on python3.  There are visual differences
between the outputs of python2 and python3 that would muddy the waters
too much if attempted to handle in the examples.  In addition, the
BOM is hard to handle in a non-distracting manner in python2, so
just run doctests in python3.